% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_task.R
\name{add_task}
\alias{add_task}
\title{Add a Task}
\usage{
add_task(name, timeline = get_timeline(), list_id = NULL, parse = FALSE,
  parent = NULL)
}
\arguments{
\item{name}{The task name.}

\item{timeline}{A timeline (see \code{get_timeline}).}

\item{list_id}{An id of a list to which the task is added.}

\item{parse}{If \code{parse = TRUE} Smart Add will be used to process the task.}

\item{parent}{(Pro feature): when a parent task id is provided, the task is
created as a sub-task.}
}
\description{
Add a Task
}

