% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authenticate_dairy}
\alias{authenticate_dairy}
\title{Authenticate Application with Remember the Milk}
\usage{
authenticate_dairy()
}
\description{
This function will authenticate the package with your online account.
A browser will be opened, in which will can authorize the dairy package.
You may need to log in to your account, if you are not already logged in.
}

