% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timeline}
\alias{get_timeline}
\title{Get or Create a new Timeline}
\usage{
get_timeline(new = FALSE)
}
\arguments{
\item{new}{logical indicating whether a new timeline should be created.}
}
\description{
A new timeline will be created if no existing timeline exists. If it
does exist and \code{new = FALSE} then the existing timeline is returned.
}

