% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.quairy_result}
\alias{print.quairy_result}
\title{Print Method for Quairy Results}
\usage{
\method{print}{quairy_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{quairy_result}}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for Quairy Results
}

