% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quairy_status.R
\name{quairy_status}
\alias{quairy_status}
\title{Check status of a Quairy Result.}
\usage{
quairy_status(quairy_result)
}
\arguments{
\item{quairy_result}{An object of type \code{quairy_result}.}
}
\value{
\code{TRUE} for "ok", \code{FALSE} otherwise.
}
\description{
The 'Remember the Milk' API method results includes a status.
This function checks whether this is "ok", or not.
}

