% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quairy.R
\name{quairy}
\alias{quairy}
\title{Low-Leveld Query Function for dairy}
\usage{
quairy(method, ..., .auth = TRUE)
}
\arguments{
\item{method}{The name of the API method to call.}

\item{...}{name-value parameter pairs.}

\item{.auth}{logical indicating whether authentication is required.}
}
\description{
This is a low-level interface to the methods in the API found
here: \url{https://www.rememberthemilk.com/services/api/}.
}

